Move forms input to a folder
