+++ src/dashboard/other/src/app.tsx
@@ -1,5 +1,5 @@
-import { CategoryNameInput } from "components/scoreboard/dashboard/category-name-input";
-import { WeekNumberInput } from "components/scoreboard/dashboard/week-number-input";
+import { CategoryNameInput } from "components/scoreboard/dashboard/inputs/category-name-input";
+import { WeekNumberInput } from "components/scoreboard/dashboard/inputs/week-number-input";
 import { useGraphicsReplicant } from "hooks/replicants/use-graphics-replicant";
 import { useReplicant } from "hooks/use-replicant";
 import { ReactElement, useEffect, useId, useState } from "react";
